10067 Remove commented lines
