+++ it/test/uk/gov/hmrc/tai/integration/TaxCodeMismatchSpec.scala
@@ -31,7 +31,7 @@ import uk.gov.hmrc.tai.model.admin.{HipToggleIabds, RtiCallToggle}
 
 import scala.concurrent.Future
 
-class TaxCodeMismatchHipToggleTaxAccountOnSpec extends IntegrationSpec {
+class TaxCodeMismatchSpec extends IntegrationSpec {
 
   override def beforeEach(): Unit = {
     super.beforeEach()
@@ -45,9 +45,6 @@ class TaxCodeMismatchHipToggleTaxAccountOnSpec extends IntegrationSpec {
     when(mockFeatureFlagService.getAsEitherT(eqTo[FeatureFlagName](RtiCallToggle))).thenReturn(
       EitherT.rightT(FeatureFlag(RtiCallToggle, isEnabled = false))
     )
-//    when(mockFeatureFlagService.get(eqTo[FeatureFlagName](HipToggleEmploymentDetails))).thenReturn(
-//      Future.successful(FeatureFlag(HipToggleTaxAccount, isEnabled = false))
-//    )
     when(mockFeatureFlagService.get(eqTo[FeatureFlagName](HipToggleIabds))).thenReturn(
       Future.successful(FeatureFlag(HipToggleIabds, isEnabled = false))
     )
