+++ it/test/uk/gov/hmrc/tai/integration/UpdateIncomeSpec.scala
@@ -33,7 +33,7 @@ import uk.gov.hmrc.tai.model.nps2.IabdType.hipMapping
 
 import scala.concurrent.Future
 
-class UpdateIncomeHipToggleTaxAccountAndIabdEmploymentOnSpec extends IntegrationSpec {
+class UpdateIncomeSpec extends IntegrationSpec {
 
   override def beforeEach(): Unit = {
     super.beforeEach()
@@ -46,10 +46,6 @@ class UpdateIncomeHipToggleTaxAccountAndIabdEmploymentOnSpec extends Integration
     when(mockFeatureFlagService.getAsEitherT(eqTo[FeatureFlagName](RtiCallToggle))).thenReturn(
       EitherT.rightT(FeatureFlag(RtiCallToggle, isEnabled = false))
     )
-
-//    when(mockFeatureFlagService.get(eqTo[FeatureFlagName](HipToggleEmploymentDetails))).thenReturn(
-//      Future.successful(FeatureFlag(HipToggleTaxAccount, isEnabled = false))
-//    )
     when(mockFeatureFlagService.get(eqTo[FeatureFlagName](HipToggleIabds))).thenReturn(
       Future.successful(FeatureFlag(HipToggleIabds, isEnabled = false))
     )
